+++ app/(task)/projects/[id]/tasks/[taskId]/page.tsx
@@ -97,7 +97,6 @@ export default async function TaskPage({ params }: TaskPageProps) {
           allTasks={allTasks}
           initialMessages={initialMessages}
           isReadonly={isReadonly}
-          snapshots={chat.snapshots ?? []}
         />
       </div>
     </div>